+++ webknossos/test.sh
@@ -4,7 +4,7 @@ set -eEuo pipefail
 source local_wk_setup.sh
 
 # Using forkserver instead of spawn is faster. Fork should never be used due to potential deadlock problems.
-export MULTIPROCESSING_DEFAULT_START_METHOD=forkserver
+export MULTIPROCESSING_DEFAULT_START_METHOD=${MULTIPROCESSING_DEFAULT_START_METHOD:-forkserver}
 
 # Note that pytest should be executed via `python -m`, since
 # this will ensure that the current directory is added to sys.path