+++ webknossos/Changelog.md
@@ -10,9 +10,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/) `MAJOR.MIN
 For upgrade instructions, please check the respective _Breaking Changes_ sections.
 
 ## Unreleased
-[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.1...HEAD)
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.2...HEAD)
 
 ### Breaking Changes
+
+### Added
+
+### Changed
+
+### Fixed
+
+
+## [2.4.2](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.2) - 2025-07-07
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v2.4.1...v2.4.2)
 
 ### Added
 - Added operations for `Project`, `Task` and `TaskType` classes and objects: [#1294](https://github.com/scalableminds/webknossos-libs/pull/1294)
@@ -33,10 +43,9 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 - Added `TaskExperience` class to store domain and value pair.
 - The libs client now uses versioned datastore routes for improved compatibility across webknossos versions. [#1328](https://github.com/scalableminds/webknossos-libs/pull/1328)
 
-### Changed
-
 ### Fixed
 - Fixed the tests for the newest WEBKNOSSOS version. [#1328](https://github.com/scalableminds/webknossos-libs/pull/1328)
+
 
 
 ## [2.4.1](https://github.com/scalableminds/webknossos-libs/releases/tag/v2.4.1) - 2025-07-03